Refactor: Consolidate utility modules
Consolidated all utility functions from clintrials/coll.py and clintrials/logging.py into the clintrials/utils.py module and removed the now-empty files.

This creates a cleaner, more organized project structure.
