+++ tests/test_coll.py
@@ -1,6 +1,6 @@
 import pytest
 
-from clintrials.coll import to_1d_list
+from clintrials.utils import to_1d_list
 
 
 @pytest.mark.parametrize(
